+++ .eslintrc.json
@@ -1,0 +1,14 @@
+{
+    "env": {
+      "es2017": true,
+      "browser": true
+    },
+    "extends": "htmlacademy/vanilla",
+    "parserOptions": {
+      "ecmaVersion": 2018,
+      "sourceType": "module"
+    },
+    "rules":  {
+    },
+    "root": true
+}
